+++ sourcespec/source_residuals.py
@@ -19,7 +19,6 @@ from glob import glob
 from collections import defaultdict
 from argparse import ArgumentParser
 import numpy as np
-from scipy.interpolate import interp1d
 import matplotlib
 import matplotlib.pyplot as plt
 from sourcespec._version import get_versions
@@ -267,9 +266,8 @@ def compute_mean_residuals(residual_dict, min_spectra=20):
             spec_interp.freq = freq_array
             # spec_slice.data must exist, so we create it as zeros
             spec_interp.data = np.zeros_like(freq_array)
-            # interpolate data_mag to the new frequency array
-            f = interp1d(spec.freq, spec.data_mag, bounds_error=False)
-            spec_interp.data_mag = f(freq_array)
+            # interpolate data and data_mag to the new frequency array
+            spec_interp.interp_data_to_new_freq(freq_array)
             # norm is 1 where interpolated data_mag is not nan, 0 otherwise
             norm = (~np.isnan(spec_interp.data_mag)).astype(float)
             # Replace nan data_mag values with zeros for summation